[CSL-1740] Ensure that fake change address is always long enough
Previous implementation was using one of output addresses as a fake
change address. But it can happen that 'getNewAddress' generates an address
which is longer than the fake change address, which in turn will lead to
tx fee being too small and invalid tx will be created.
Note that this problem is quite important already, even though we can only
create bootstrap era addresses with HD payload. For more details about it
please see my comment in CSL-1740.

The solution is a bit hacky, but it seems there is no better way to resolve
problem. 'getFakeChangeAddress' was added to 'MonadAddresses' which must
create an address with size greater than or equal to maximal possible size of
address generated by 'getNewAddress'. An address generated by
'getFakeChangeAddress' is used to create a fake transaction to compute fee.
In rare cases fee can be slightly greater than limit (by about 1-3 bytes
multiplied by ratio), but it's quite negligible.
Unfortunately we can't generate a real change address before computing fee,
because it requires a passphrase, but we don't want to require passphrase to
compute fee.